+++ client/src/app/router.jsx
@@ -4,6 +4,7 @@ import AutomationsPage from '../pages/AutomationsPage'
 import InboxPage from '../pages/InboxPage'
 import LeadsPage from '../pages/LeadsPage'
 import NotFoundPage from '../pages/NotFoundPage'
+import PaymentsPage from '../pages/PaymentsPage'
 
 export const router = createBrowserRouter([
   {
@@ -25,6 +26,10 @@ export const router = createBrowserRouter([
       {
         path: 'automations',
         element: <AutomationsPage />,
+      },
+      {
+        path: 'payments',
+        element: <PaymentsPage />,
       },
       {
         path: '*',